refactor: seats index export, overall css trimmed
