+++ src/components/seat/X/index.ts
@@ -1,0 +1,2 @@
+export * from './CX1Cluster';
+export * from './CX2Cluster';
