+++ src/components/utils/Divider.tsx
@@ -1,3 +1,3 @@
-export default function Divider() {
+export function Divider() {
   return <hr className="w-full border-[1px] border-darkblue" />;
 }
