+++ src/components/utils/ProfileSkeleton.tsx
@@ -1,6 +1,6 @@
 import { Skeleton } from '@/components/ui/skeleton';
 
-export default function ProfileSkeleton() {
+export function ProfileSkeleton() {
   return (
     <div className="flex items-center gap-2 p-6 pt-2">
       <Skeleton className="size-16 rounded-full md:size-28" />
