+++ src/pages/404.tsx
@@ -5,12 +5,12 @@ import logo from '@/assets/logo/logo.svg';
 
 export default function NotFoundPage() {
   return (
-    <main className="flex min-h-[100vh] w-full flex-col text-darkblue sm:flex-row">
+    <main className="flex min-h-svh w-full flex-col text-darkblue sm:flex-row">
       <div className="pt-8 sm:relative sm:min-h-[50vh] sm:w-1/2">
         <Image src="/image/fill/left.png" alt="left" layout="fill" className="hidden sm:block" />
       </div>
       <div className="w-full bg-white sm:w-1/2">
-        <div className="flex h-full min-h-[50vh] flex-col items-center justify-center gap-8 sm:gap-12">
+        <div className="flex h-full min-h-[50svh] flex-col items-center justify-center gap-8 sm:gap-12">
           <Image src={logo} alt="logo" priority className="h-28 w-fit" />
           <span className="flex flex-col items-center gap-4">
             <h2 className=" text-2xl sm:text-3xl">이 페이지는 존재하지 않습니다.</h2>
@@ -22,7 +22,7 @@ export default function NotFoundPage() {
           </Link>
         </div>
       </div>
-      <div className="relative flex min-h-[50vh] w-full flex-col items-center justify-center sm:hidden">
+      <div className="relative flex min-h-[50svh] w-full flex-col items-center justify-center sm:hidden">
         <Image src="/image/fill/bottom.png" alt="bottom" layout="fill" />
       </div>
     </main>
